+++ src/treeML.h
@@ -1,4 +1,4 @@
 #define PHYML
 
-int treeML(const struct aln *alignment[], char** treeString, double* kappa);
+int treeML(const struct aln *alignment[], char** treeString, float* kappa);
