+++ src/treeSimulate.h
@@ -5,8 +5,8 @@ TNode* getLCA(TTree* tree, TNode* nodeA, TNode* nodeB);
 void tree2aln(TTree* tree, struct aln *alignment[]);
 TTree* string2tree(char* treeString);
 
-void simulateTree(TTree* tree, double freqs[], double kap, int L);
+void simulateTree(TTree* tree, float freqs[], float kap, int L);
 
-double** getDistanceMatrix(TTree* tree, struct aln *alignment[]);
+float** getDistanceMatrix(TTree* tree, struct aln *alignment[]);
 
 void freeSeqgenTree(TTree* tree);
